Split the dynamic_honey_badger module.
