+++ src/dynamic_honey_badger/change.rs
@@ -1,0 +1,33 @@
+use crypto::PublicKey;
+
+/// A node change action: adding or removing a node.
+#[derive(Clone, Eq, PartialEq, Serialize, Deserialize, Hash, Debug)]
+pub enum Change<NodeUid> {
+    /// Add a node. The public key is used only temporarily, for key generation.
+    Add(NodeUid, PublicKey),
+    /// Remove a node.
+    Remove(NodeUid),
+}
+
+impl<NodeUid> Change<NodeUid> {
+    /// Returns the ID of the current candidate for being added, if any.
+    pub fn candidate(&self) -> Option<&NodeUid> {
+        match *self {
+            Change::Add(ref id, _) => Some(id),
+            Change::Remove(_) => None,
+        }
+    }
+}
+
+/// A change status: whether a node addition or removal is currently in progress or completed.
+#[derive(Clone, Eq, PartialEq, Serialize, Deserialize, Hash, Debug)]
+pub enum ChangeState<NodeUid> {
+    /// No node is currently being considered for addition or removal.
+    None,
+    /// A change is currently in progress. If it is an addition, all broadcast messages must be
+    /// sent to the new node, too.
+    InProgress(Change<NodeUid>),
+    /// A change has been completed in this epoch. From the next epoch on, the new composition of
+    /// the network will perform the consensus process.
+    Complete(Change<NodeUid>),
+}
