+++ src/dynamic_honey_badger/err.rs
@@ -1,0 +1,17 @@
+use bincode;
+
+use honey_badger;
+
+error_chain!{
+    links {
+        HoneyBadger(honey_badger::Error, honey_badger::ErrorKind);
+    }
+
+    foreign_links {
+        Bincode(Box<bincode::ErrorKind>);
+    }
+
+    errors {
+        UnknownSender
+    }
+}
